Move gallery id to metadata
Pulled the gallery id from the title and impl the Number field as per spec.  Komga will import it as the series number (so it is still in front visually) and it is added as a tag for searchability.  LANraragi users can opt to use a tag rule to omit the nhentai prefix.  A better impl may be possible but for now I did this attempt adding an additional parameter to filter_and_combine_tags as an Option which allows it to be set to None for those calls not using it.