Add raw timestamp to ComicInfo generation
This is mostly a personal patch for ComicInfo import.  ComicInfo on LANraragi had to be modified as well to locate and use this data.  This made it so I didn't have to try and convert M/D/Y info in the file back to epoch for display in the GUI.  It is set in the timestamp tag for each item.